// src/app/models/document-tracking-models/document-tracking/dtos/document-tracking-dto.ts
export interface DocumentTrackingDto {
  id?: number|null;
  documentTrackingCode?: string|null;
  referenceNumber?: string|null;
  personEnglishName?: string|null;
  personArabicName?: string|null;
  personShortName?: string|null;
  personId?: number|null;
  registerIdc?: string|null;
  registerIdn?: number|null;
  description?: string|null;
  registerCode?: string|null;
}